requirements.txt
pandas
numpy
scikit-learn
dvc
mlflow
pyyaml
